requirements.txt
streamlit
pandas
openpyxl
google-cloud-aiplatform
pyyaml
